Implement clear command (#73)
* Implement Clear Command

* Fix Parser

* Fix issue with clear of completed task and events

* Fix Mark command

* Fix parser issue

* Fix Merge conflict with master

* Add Clear Command test cases.

* Fix FindCommandTest case with modified clear command

* Add Clear Command test heuristics and fix minor errors

* Add author's name to relevant commands, parsers and Junit tests
Add test heuristics for delete and mark command
Add utility function in Commandtest
Fix test case for delete command

* Fix Message_Usage for clear command

* Modified Clear Command's execution, removed null case return

+++ src/main/java/seedu/task/logic/parser/DeleteParser.java
@@ -12,6 +12,10 @@ import seedu.task.logic.commands.Command;
 import seedu.task.logic.commands.DeleteCommand;
 import seedu.task.logic.commands.IncorrectCommand;
 
+/**
+ * Responsible for validating and preparing the arguments for DeleteCommand execution
+ * @author Tiankai
+ */
 public class DeleteParser implements Parser {
 
     public DeleteParser() {}
@@ -36,6 +40,8 @@ public class DeleteParser implements Parser {
         final Matcher eventMatcher = EVENT_DATA_ARGS_FORMAT.matcher(args.trim());
         if (taskMatcher.matches()) {
             int index = Integer.parseInt(taskMatcher.group("index"));
+            if (index==0) return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+            
             try {
                 return new DeleteTaskCommand(index);
             } catch (NumberFormatException ive) {
@@ -43,6 +49,8 @@ public class DeleteParser implements Parser {
             }
         } else if (eventMatcher.matches()){
             int index = Integer.parseInt(eventMatcher.group("index"));
+            if (index==0) return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+            
             try {
                 return new DeleteEventCommand(index);
             } catch (NumberFormatException ive) {
